Move all tree view out of contrib Fixes #110374
