+++ src/vs/workbench/browser/parts/views/media/views.css
